vsock_proxy: rename starter.rs
Rename starter.rs module to proxy.rs to better reflect its purpose.
